+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-09-08 02:59:39
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 986105ca69203e69cd31250a1b5530259b6af8e6
+
+## [ForwardWidgets-2kuai] - 2025-09-08 02:59:39
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: bdac42eb704af8c2d8aac61b6fb05e44c11068a9
+
+## [ForwardWidgets-pack1r] - 2025-09-08 02:59:39
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 15897affdf5e60159ec2b47b3effe3e004c5ca77
+
+
 ## [ForwardWidgets-huangxd] - 2025-09-07 02:56:11
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main